t - unit tested styles for overall score section for tennis game 1
