+++ angular/src/app/tennis-game2/tennis-game2.component.spec.ts
@@ -1,13 +1,10 @@
 import {async} from '@angular/core/testing';
 import {TennisGame2Component} from './tennis-game2.component';
 import {
-  expectedTennisScores,
-  getScoreButton,
-  overallScoreLabel,
-  player1ScoreInput,
-  player2ScoreInput,
   TennisComponentTester
-} from '../../test/testHelper';
+} from '../../test/tennisTester';
+import {getScoreButton, overallScoreLabel, player1ScoreInput, player2ScoreInput} from '../../test/selectors';
+import {expectedTennisScores} from '../../test/expectedResults';
 
 describe('Tennis Game 2', () => {
   let tennisTester: TennisComponentTester;
