+++ angular/src/test/expectedStyles.ts
@@ -1,0 +1,6 @@
+export const tennisBallOpticYellowColor = 'rgb(204, 255, 0)';
+export const slightlyRoundedBottomCorners = '0px 0px 5px 5px';
+export const zeroPixels = '0px';
+export const twentyPixels = '20px';
+export const flex = 'flex';
+export const spaceBetween = 'space-between';
