+++ angular/src/test/selectors.ts
@@ -1,0 +1,5 @@
+export const overallScoreLabel = '#overall-score';
+export const overallScoreSection = '#overall-score-section';
+export const player1ScoreInput = '#player-one-score';
+export const player2ScoreInput = '#player-two-score';
+export const getScoreButton = '#get-score-button';
